Edit Droid Validation and Styling work

+++ database/migrations/2021_10_27_180639_update_events_table_with_contact_id.php
@@ -14,7 +14,7 @@ class UpdateEventsTableWithContactId extends Migration
     public function up()
     {
         Schema::table('events', function (Blueprint $table) {
-            $table->integer('venue_contacts_id');
+            $table->unsignedBigInteger('venue_contacts_id');
 
             // Foreign
             $table->foreign('venue_contacts_id')->references('id')->on('venue_contacts')->nullable();
